+++ resources/views/kolekcia.blade.php
@@ -10,6 +10,9 @@
 <meta property="og:url" content="{!! Request::url() !!}" />
 <meta property="og:image" content="{!! URL::to($collection->getHeaderImage()) !!}" />
 <meta property="og:site_name" content="web umenia" />
+@foreach ($collection->ContentImages() as $image )
+    <meta property="og:image" content="{!! $image !!}" />
+@endforeach
 @stop
 
 @section('title')