+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file[^1].
 - preview main / header image within article and collection admin editor
 - prevent storing of empty alternative locales for articles / collections
 - small locale fixes
-
+- absolute urls for og:image
 
 ## [2.6.1] - 2020-05-04
 ### Fixed